Fix: rerender custom actions on reopen side panel #60
The cause of this error in **not** in the components library. It was (is) an implementation issue: when createing a new instance of a form model, with custom actions defined, the new instance need to have the custom actions also be defined.
If a custom action is null, the corresponding button will not be rendered.